+str reintroduced the TCK, 0.4.0.M2-SNAPSHOT, passing all tests
Cheers from JavaZone!

+++ akka-stream/src/main/scala/akka/stream/impl/Messages.scala
@@ -12,7 +12,7 @@ private[akka] case object SubscribePending
 /**
  * INTERNAL API
  */
-private[akka] case class RequestMore(subscription: ActorSubscription[_], demand: Int)
+private[akka] case class RequestMore(subscription: ActorSubscription[_], demand: Long)
 /**
  * INTERNAL API
  */
